wow, so... getting the animation to play once was way easy.  Getting the animation to play more than once is still not complete.  I made it play by attaching a css class which contained the animation.  My plan was to then remove the class, but javascript's timer operations were being dicks.  Essentially, I always had to remove it immediately, and then the animation wouldn't have any time to play.
Thus, I am doing the following Yak-shaving:

I'm in the middle of changing control schemes so that it uses arrow keys.  I just made it so that the selection moves up and down when you press up and down.  Next is getting a css class for selected.  Then making it so the space bar either selects the current technique if it is the player turn OR makes the enemy attack if it is the enemy's turn.  This separation will then allow time to take off the css classes before putting them back on again.

Next steps:
finish changing control scheme
make sure animations work now
use stats in damage calculation
AJAX in from monsters
refactor tests